Add support for multiple containers (#9)
* Start adding multiple containers support

* Improve logging

* Prevent ws4py logs propagation

* Use a creation counter to ensure container name consistence

When creating containers without speciying their names, we should have a mechanism allowing us to identify each container before it is actually created using LXD. This
is why a creation counter has been added to the Container class. This attribute (that is in the namespace of the Container class) will be used to generate container
names when they are not provided by the .nomad.yml config file.

* Fix the way containers are identified in the scope of the project directory

Prior to this commit, a container was identified using its homedir. This do not allow us to support multi containers. So a property has been added in order to generate
"nomad IDs" used to uniquely identify each container in the scope of the project directory.

* Display the help when the action is not specified in the command line

* Add a way to execute each action for a specific container

* Improve container config generation