Improve container names for multi-containers (#12)
* Add a way to get "deterministic" container names

Prior to this commit, the LXD names of the containers were generated when each container was created. Thus, we did not know what was the LXD name of each
container in the "nomad" scope. This forced us to add things like "NID" in order to discover a specific container by iterating over each container managed by LXD
and checking its "NID". This was not the most effective solution.

This commit introduces a "LXD name" that is computed for each container by using a unique identifier. This identifier is stored inside a hidden metadata file at
the root of the Nomad project. This allows to know the LXD names (which are unique) of the containers before they are started. By doing this we can use the
Client.containers.get method provided by pylxd in order to easily retrieve containers by using their names.

* Fix incorrect file modes

* Add a way to use lXD names without hidden metadata files

* Fix 'shell' action

* Remove the need to slugify project names by enforcing their presence in config files

* Enforce the use of container names