Fix color and logger scope of the shell command
I noticed that the shell command's error when the container wasn't
running wasn't properly colored, so I added a proper check (the same as
in the provision command).

Then, I noticed that the logger wasn't properly "container scoped" in
that same shell command because the call wasn't wrapped in the same
fancy generator that wraps a logger around our call, so I fixed this.