docs: add "Problems?" section to "Getting Started" (#35)
The old readme had something about trying privileged mode when
encountering problems and this disappeared in the new docs. Considering
that `debian/jessie` is a popular image and that it doesn't work in
unprivileged mode, I think it's important to have that section to avoid
being swamped with troubleshooting issues of this type on github.