shell: add a way to set environment variables (#50)
* shell: add a way to set environment variables

Add a new config dictionary in the `shell` key: `environment`. All
values in this dictionary are passed to LXD's config under the
"environment.*" config type.

We set this config when `shell` is called, which is a bit at odds with
the rest of the container config, which is usually set at creation time.
We do this so that a user who fiddles with her `lxdock.yml` file doesn't
have to restart/recreate her container to benefit from the change she's
just made. They will be effective on the next `lxdock shell` call.

* Change shell-env test's container name in test_container

It apparently prevents it from passing.

* shell-env: coerce env values into strings

Otherwise, we can set env variables to values like "42" without
explicitly quoting them in our yaml.

* Widen the scope of the "environment" config

Rather than being specific to `shell`, `environment` is now a top-level
configuration and it now affects both `provision` and `shell`.