Restore ci-base-setup.sh and remove from Makefile
It seems all that was missing from this script was "sudo snap start lxd", without this some tests would fail before but this fixes it.

Also create vagrant-base-setup.sh rather than have this in the Vagrantfile, since I've added some lines to this.

vagrant-base-setup.sh installs python3-virtualenv because that isn't available on Trusty and Travis.