Fix shell -c command is now like Vagrant
The shell command feature did not seem to work when running two commands because it was using shlex.quote on each argument.

See: issue #129

With the current syntax I was unable to work out how to simply run two commands in a container, Vagrant uses this much simpler syntax which makes much more sense:

    vagrant ssh -c "command1; command2"

Also Vagrant will produce an error if you add another argument so this will not work:

    vagrant ssh -c command1 command2

It seemed lxdock followed something like that except it auto quoted each argument, it was quite confusing and I couldn't figure out how two run two commands and that produced an error.

I have changed lxdock to be the same as how vagrant works:

    lxdock shell -c "command1; command2"

Fixed tests and updated zsh config, the bash one doesn't need updating by the looks of it.

Closes #129