Correctly set the SHELL variable
Right now, since we are launching shell commands by first transferring a script and then launching the script as if it is a shell (see man su for the -s option), it sets the SHELL environment variable to the script.

Some applications relies on using the SHELL environment variable to launch subprocesses, which could result in a forkbomb. This PR fixes this behaviour by setting the SHELL to /bin/sh, which is accurate.

Based on original PR#153 by shuhaowu, checking to see if we get two failed tests agains latest master.

In addition, fixed two tests that were failing as part of this change, which was not obvious before as CI was broken at the time.